backend/api/requirements.txt
flask_sqlalchemy
flask
flask_restful
requests
python-dotenv
sqlalchemy
flask-migrate
alembic
psycopg2
google-auth